+++ app/src/main/java/com/example/kouveepetshop/Pengelolaan/Supplier/Supplier_Edit.java
@@ -7,6 +7,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -34,6 +35,8 @@ public class Supplier_Edit extends AppCompatActivity {
 
     private SharedPrefManager sharedPrefManager;
 
+    private boolean doubleClickDelete = false;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -46,32 +49,48 @@ public class Supplier_Edit extends AppCompatActivity {
         edit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                editSupplier();
+                if (validasi()) {
+                    editSupplier();
 
-                new Handler().postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        Intent returnIntent = new Intent();
-                        setResult(RESULT_OK,returnIntent);
-                        finish();
-                    }
-                }, 500);
+                    new Handler().postDelayed(new Runnable() {
+                        @Override
+                        public void run() {
+                            Intent returnIntent = new Intent();
+                            setResult(RESULT_OK, returnIntent);
+                            finish();
+                        }
+                    }, 500);
+                }
             }
         });
 
         delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                deleteSupplier();
+                if (doubleClickDelete) {
+                    deleteSupplier();
 
-                new Handler().postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        Intent returnIntent = new Intent();
-                        setResult(RESULT_OK,returnIntent);
-                        finish();
-                    }
-                }, 500);
+                    new Handler().postDelayed(new Runnable() {
+                        @Override
+                        public void run() {
+                            Intent returnIntent = new Intent();
+                            setResult(RESULT_OK, returnIntent);
+                            finish();
+                        }
+                    }, 500);
+                }
+                else {
+                    doubleClickDelete = true;
+                    Toast.makeText(Supplier_Edit.this, "Tekan Lagi Untuk Delete", Toast.LENGTH_SHORT).show();
+
+                    new Handler().postDelayed(new Runnable() {
+
+                        @Override
+                        public void run() {
+                            doubleClickDelete = false;
+                        }
+                    }, 2000);
+                }
             }
         });
     }
@@ -170,5 +189,60 @@ public class Supplier_Edit extends AppCompatActivity {
         kota_text       = findViewById(R.id.supplier_edit_kota);
         edit            = findViewById(R.id.supplier_edit_edit);
         delete          = findViewById(R.id.supplier_edit_delete);
+    }
+
+    private boolean validasi() {
+        int cek = 0;
+        if (nama_text.getText().toString().equals("")) {
+            nama_text.setError("Nama Tidak Boleh Kosong");
+            cek = 1;
+        }
+        else if (nama_text.getText().toString().length() < 3) {
+            nama_text.setError("Panjang Nama Minimal 3 Karekter");
+            cek = 1;
+        }
+
+        else if (!nama_text.getText().toString().matches("[a-zA-Z ]+")) {
+            nama_text.setError("Format Nama Salah");
+            cek = 1;
+        }
+
+        if (no_telp_text.getText().toString().equals("")) {
+            no_telp_text.setError("Nomor Telepon Tidak Boleh Kosong");
+            cek = 1;
+        }
+        else if (no_telp_text.getText().toString().length() < 10 || no_telp_text.getText().toString().length() > 13 ) {
+            no_telp_text.setError("Nomor Telepon 10 - 13 Karakter");
+            cek = 1;
+        }
+
+        else if (!String.valueOf(no_telp_text.getText().toString().charAt(0)).equals("0") && !String.valueOf(no_telp_text.getText().toString().charAt(1)).equals("8")) {
+            no_telp_text.setError("Format Nomor Telepon Salah");
+            cek = 1;
+        }
+
+        if (alamat_text.getText().toString().equals("")) {
+            alamat_text.setError("Alamat Tidak Boleh Kosong");
+            cek = 1;
+        }
+        else if (alamat_text.getText().toString().length() < 3) {
+            alamat_text.setError("Panjang Alamat Minimal 3 Karekter");
+            cek = 1;
+        }
+
+        if (kota_text.getText().toString().equals("")) {
+            kota_text.setError("Kota Tidak Boleh Kosong");
+            cek = 1;
+        }
+        else if (kota_text.getText().toString().length() < 3) {
+            kota_text.setError("Panjang Kota Minimal 3 Karekter");
+            cek = 1;
+        }
+
+        else if (!kota_text.getText().toString().matches("[a-zA-Z ]+")) {
+            kota_text.setError("Format Kota Salah");
+            cek = 1;
+        }
+        return cek == 0;
     }
 }
